Move useSwapButtonContext
