+++ src/custom/pages/NewSwap/index.tsx
@@ -25,7 +25,7 @@ import { useWrapType, WrapType } from 'hooks/useWrapCallback'
 import { getInputReceiveAmountInfo, getOutputReceiveAmountInfo } from 'pages/NewSwap/helpers/tradeReceiveAmount'
 import { useSwapFlowContext } from 'pages/Swap/swapFlow/useSwapFlowContext'
 import { SwapButton } from 'pages/Swap/components/SwapButton/SwapButton'
-import { useSwapButtonContext } from 'pages/NewSwap/hooks/useSwapButtonContext'
+import { useSwapButtonContext } from 'pages/Swap/hooks/useSwapButtonContext'
 import { OperationType } from 'components/TransactionConfirmationModal'
 import { useModalIsOpen, useOpenModal } from 'state/application/hooks'
 import { ApplicationModal } from '@src/state/application/reducer'
