+++ src/custom/pages/Swap/hooks/useSwapButtonContext.ts
@@ -1,25 +1,25 @@
 import { useWeb3React } from '@web3-react/core'
 import { useWalletInfo } from 'hooks/useWalletInfo'
-import { useDerivedSwapInfo, useDetectNativeToken, useSwapState } from 'state/swap/hooks'
+import { useDerivedSwapInfo, useDetectNativeToken, useSwapState } from '@src/state/swap/hooks'
 import useTransactionDeadline from '@src/hooks/useTransactionDeadline'
 import { useExpertModeManager } from '@src/state/user/hooks'
-import { useCloseModals, useToggleWalletModal } from 'state/application/hooks'
+import { useCloseModals, useToggleWalletModal } from '@src/state/application/hooks'
 import { useSwapConfirmManager } from 'pages/Swap/hooks/useSwapConfirmManager'
 import { Field } from '@src/state/swap/actions'
 import { TradeType } from '@uniswap/sdk-core'
-import { computeSlippageAdjustedAmounts } from 'utils/prices'
-import { useWrapType, useWrapUnwrapError } from 'hooks/useWrapCallback'
+import { computeSlippageAdjustedAmounts } from '@src/utils/prices'
+import { useWrapType, useWrapUnwrapError } from '@src/hooks/useWrapCallback'
 import { useCallback } from 'react'
 import { logSwapFlow } from 'pages/Swap/swapFlow/logger'
 import { swapFlow } from 'pages/Swap/swapFlow'
-import { useApproveCallbackFromTrade } from 'hooks/useApproveCallback'
+import { useApproveCallbackFromTrade } from '@src/hooks/useApproveCallback'
 import { useGnosisSafeInfo } from 'hooks/useGnosisSafeInfo'
-import { useIsSwapUnsupported } from 'hooks/useIsSwapUnsupported'
+import { useIsSwapUnsupported } from '@src/hooks/useIsSwapUnsupported'
 import { ApproveButtonProps } from 'pages/Swap/components/ApproveButton'
 import { getSwapButtonState } from 'pages/Swap/helpers/getSwapButtonState'
 import { SwapButtonContext } from 'pages/Swap/components/SwapButton/SwapButton'
 import { useGetQuoteAndStatus } from 'state/price/hooks'
-import { OperationType } from 'components/TransactionConfirmationModal'
+import { OperationType } from '@src/components/TransactionConfirmationModal'
 import { SwapFlowContext } from 'pages/Swap/swapFlow/types'
 
 export interface SwapButtonInput {
